Backend/error handling

+++ backend/src/main/java/com/kanban/backend/service/UserService.java
@@ -6,6 +6,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.regex.Pattern;
 
 @Service
 @RequiredArgsConstructor
@@ -21,10 +22,33 @@ public class UserService {
     }
 
     public User addUser(User user) {
+        if (userRepository.existsByEmail(user.getEmail())) {
+            return null;
+        }
+
+        String email = user.getEmail();
+        String emailRegex = "^(?=.{1,64}@)[A-Za-z0-9_-]+(\\.[A-Za-z0-9_-]+)*@"
+                + "[^-][A-Za-z0-9-]+(\\.[A-Za-z0-9-]+)*(\\.[A-Za-z]{2,})$";
+
+        boolean isEmailValid = Pattern
+                .compile(emailRegex)
+                .matcher(email)
+                .matches();
+
+        if (!isEmailValid) {
+            return null;
+        }
+
         return userRepository.save(user);
     }
 
-    public void deleteUserById(Long id) {
-        userRepository.deleteById(id);
+    public boolean deleteUserById(Long id) {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+
+            return true;
+        }
+
+        return false;
     }
 }
